move clone function into utils.py
