+++ .github/CODEOWNERS
@@ -433,6 +433,8 @@
 /jobs/base_model_validator_test.py @brianrodri
 /jobs/base_model_validator_errors.py @brianrodri
 /jobs/base_model_validator_errors_test.py @brianrodri
+/jobs/utils.py @brianrodri
+/jobs/utils_test.py @brianrodri
 /core/controllers/cron*.py @nithusha21
 /core/domain/cron_services*.py @nithusha21
 /main_cron.py @nithusha21
